Update patch set 1
Patch Set 1: Code-Review+1

This looks very useful.

It's good that this plugin is available under the Apache 2.0 license, but I think we should document somewhere that we use this plugin under this license.

The runtime dependencies are listed with their licenses within the app in the help screen. I guess it's not needed to expose this dependency on that level, but maybe we should add a section for this to the README file.

Patch-set: 1
Label: Code-Review=+1