Update patch set 1
Patch Set 1:

> > It's good that this plugin is available under the Apache 2.0
 > > license, but I think we should document somewhere that we use
 > this
 > > plugin under this license.
 > 
 > Ok, but why do you believe we should document the use of this
 > plugin explicitly, but not e.g. the use of app dependencies?

I think we should document all dependencies.
The app dependencies are already explicitly documented as part of the help screen that the user can open in the app.

 > I
 > mean, where do you see the difference between build system
 > dependencies (like this plugin) and compile dependencies of the
 > app?

The build system dependencies are not part of the apk that people install on their phone. This is why I think they don't need to be mentioned on the help screen of the app. But still somewhere we should document them. So the README might be a good place for this. When we do this it makes sense to list the app dependencies there as well (even though it means that they are in 2 places then).

Patch-set: 1