Update patch set 1
Patch Set 1:

> I think we should document all dependencies.

For what purpose would you like to document them? If it was for developers to show what we're using, they would know that the build system / Gradle is the primary place to look at. If it was for license compliance checks, we could probably use yet another plugin like https://github.com/hierynomus/license-gradle-plugin.

 > The app dependencies are already explicitly documented as part of
 > the help screen that the user can open in the app.

Ok, I didn't check that yet.

 > The build system dependencies are not part of the apk that people
 > install on their phone. This is why I think they don't need to be
 > mentioned on the help screen of the app. But still somewhere we
 > should document them. So the README might be a good place for this.
 > When we do this it makes sense to list the app dependencies there
 > as well (even though it means that they are in 2 places then).

I've nothing against documenting the dependencies, but manually keeping a README in sync with reality just does not fly IMO. I'm suggesting to do this in a more automated way, for example by using the plugin mentioned above to generate a file at build time that gets used in the app's help screen.

Patch-set: 1