Update patch set 1
Patch Set 1:

Documenting the dependencies with their licenses is mostly for legal reasons. We want to be sure that we are compliant with the licenses.

If there is a plugin to manage this, that would be much better :-)
I totally agree that keeping track of them manually is a pain.

Patch-set: 1