Update patch set 1
Patch Set 1:

> Or did you mean asking him to actually implement it?
 > 
 > [1] https://groups.google.com/d/msg/repo-discuss/VaSJGTPUpd4/t5V6NaytCgAJ

Yes, I was meaning to ask him to implement it, mainly because I though he's the only one having the necessary rights to do so. But I'd also be happy to do it as I have plenty of experience with Android CI builds on Jenkins. I'll reply to that thread.

Patch-set: 1