Set minSdkVersion to 17 (Android 4.2)
Android 4.2 supports nested fragments [1]. This will allow us to use a
ViewPager within a Fragment.

At the moment 10% of the Android devices run with an SDK version >= 14
and <= 17 [2]. Considering that the app is only used by developers and
developers tend to have new devices with latest updates it should be
OK to increase the minSdkVersion to 17 and thus removing support for
this 10% of Android devices.

[1] http://developer.android.com/about/versions/android-4.2.html#NestedFragments
[2] http://developer.android.com/about/dashboards/index.html#Platform

Change-Id: Ic08b5761ec961ec6cbb92901908067f64a149d01
Signed-off-by: Edwin Kempin <ekempin@google.com>