Use SlidingTabLayout for tab strip on PagedChangeDetailsFragment
SlidingTabLayout layouts all tabs next to each other while
PagerTabStrip centers the selected tab and shows the previous/next tab
left/right aligned at the border of the screen. This means with
PagerTabStrip not all tabs are shown, e.g. if there are 3 tabs and the
first tab is selected the third tab is not shown. In addition long tab
titles, like 'Commit Message' may be truncated when they are shown as
previous/next tab. Because of this using SlidingTabLayout is better.

The classes SlidingTabLayout and SlidingTabStrip are copied from the
SlidingTabsBasic example project that was published by Google [1].
Both classes are available under the Apache 2 license so that we can
just use them.

As they are, SlidingTabLayout and SlidingTabStrip, cannot be styled to
look exactly the same way as the PagerTabStrip that we used so far.
The styling will be fix in the follow-up commit, so that this commit
contains the unmodified SlidingTabLayout and SlidingTabStrip classes.

[1] http://developer.android.com/samples/SlidingTabsBasic/project.html

Change-Id: I5d8d5ff89403717af8e6fd6669ac65ce6f98e9fc
Signed-off-by: Edwin Kempin <ekempin@google.com>