Fix NPE's when accessing ReviewItApp from activity
getActivity() in Fragment may return null when the fragment is not
added to an activity. This leads to NPE's when trying to get the
ReviewItApp from the activity from a callback that may be executed
after the fragment was removed from the activity. Use isAdded() to
skip the callback code when the fragment is not added to an activity.

Change-Id: I4f28660987730ff866cc12dbb76920e34b13a658
Signed-off-by: Edwin Kempin <ekempin@google.com>