Paged change details fragment: Add tab to reply
The PostReviewView already implements the controls for replying to a
change, however for displaying it in a tab the layout needs to be a
bit different (e.g. the panel with the publish buttons should be shown
at the bottom and not at the top). In addition when it is displayed in
a tab there is no need to show the commit message and the approvals as
there are extra tabs for this and it's easy to switch back and forth
between tabs. To support both layouts PostReviewView now provides 2
variants of itself.

Change-Id: Id99e18a5503e3707758d49dc18c868d5a3250576
Signed-off-by: Edwin Kempin <ekempin@google.com>