Paged unified diff view: Fix formatting of headers
When displaying the diff for each file in an own view and then paging
through the diffs, the header for the next diff was not shown if the
current diff was not fully scrolled to the top. This was caused by an
unexpected order of layout callbacks, due to reusing the
ScrollWithHeadingsView for all diffs. Fix this by not reusing the
ScrollWithHeadingsView, but instead create a new instance of it
whenever the user navigates to the next/previous diff.

Change-Id: Ic76461a7eb3cd1a547bc39eb4b4e0ea17a67fd8f
Signed-off-by: Edwin Kempin <ekempin@google.com>