Include app name and version into User-Agent header
The User-Agent header that the client sends is included in the HTTP
request log on server side. Having the app name and the version
included in the User-Agent header allows us to see from the logs how
much the app is used.

To modify the User-Agent header we need a dependency to the Apache
HTTP client. According to [1] we must add org.apache.http.legacy as
used library since the app has sdk 23 as target.

[1] http://stackoverflow.com/questions/26024908/apache-httpclient-android-gradle

Change-Id: Ie2f4dccfa1876986a98f3e985ab3624153db7ad5
Signed-off-by: Edwin Kempin <ekempin@google.com>